Análise Horizontal

Essa função calcula a análise horizontal de contas específicas.

ind_analiseHorizontal(
  dt = dt_contabil[, 1:5],
  colCod = "empresa",
  cod = "alpha",
  colTempo = "ano"
)

Arguments

dt

O banco de dados com as colunas referentes ao nome da empresa, período de análise e às contas a serem analisadas

colCod

String com o nome da coluna que se refere à coluna em que estão as empresas

cod

Vetor do tipo caracter com o código das empresas que deseja analisar

colTempo

String com o nome da coluna que se refere à coluna em que esta o tempo de análise

Details

Apresenta como resultado um data.frame com as colunas do tipo numérica com as suas respectivas análises horizontais de um período para o outro.

Informações adicionais sobre como usar o pacote, orientamos acessar o menu cntdd do Blog do Projeto contabiliDados: https://contabilidados.com.br/. Ao acessar, fazer busca pelo nome da funcao ind_analiseVertical

Contatos pelo email do Projeto contabiliDados: Email: contabilidados@ufersa.edu.br Siga-nos no Instagram: https://www.instagram.com/contabilidados @contabilidados

Examples

library(cntdd)

ind_analiseHorizontal(
dt_contabil[, 1:5],
colCod = "empresa",
cod = c("alpha", "beta"),
colTempo = "ano"
)
#> # A tibble: 12 × 5
#> # Groups:   ticker [2]
#>    ticker   ano caixaEquiv aplicFinanc clientesCP
#>    <chr>  <int>      <dbl>       <dbl>      <dbl>
#>  1 alpha   2016     NA         NA         NA     
#>  2 alpha   2017      1.02       0.378      0.0687
#>  3 alpha   2018     -0.163     -0.307      0.136 
#>  4 alpha   2019      0.624      0.161      0.0802
#>  5 alpha   2020      1.77       0.982      0.448 
#>  6 alpha   2021     -0.119     -0.562      0.320 
#>  7 beta    2016     NA         NA         NA     
#>  8 beta    2017     -0.334      0.104      0.137 
#>  9 beta    2018      0.869      0.0640     0.219 
#> 10 beta    2019      0.230      0.583      0.141 
#> 11 beta    2020      2.10      -0.272     -0.141 
#> 12 beta    2021     -0.409     -0.201      0.403