Análise Vertical e Horizontal

Essa função calcula a análise vertical e horizontal de contas específicas de um data.frame.

ind_analiseVH(
  .dt = dt_contabil,
  .contas = c("caixaEquiv", "clientesCP"),
  .contaReferenciaAV = ativoTotal,
  .colEmpresa = empresa,
  .colPeriodo = ano,
  .output = "AVAH"
)

Arguments

.dt

data.frame com, pelo menos, 4 colunas para preenchimento dos demais argumentos desta função

.contas

Vetor tipo character com o nome das colunas referente às contas de interesse na análise

.contaReferenciaAV

Nome da coluna com a conta de referência para a análise vertical

.colEmpresa

Valor Nome da coluna que identifica a empresa (pode haver mais de uma empresa no banco de dados)

.colPeriodo

Valor Nome da coluna que identifica o período da análise (pelo menos 2, por empresa, para a análise horizontal)

.output

Informar full para o banco de dados completo, AV para mostrar apenas as análises verticais, AH para mostrar apenas as análises horizontais, AHAV para mostrar, respectivamente, a análise horizontal e vertical ou AVAH para mostrar, respectivamente, a análise vertical e horizontal

Details

O usuário deve dispor de uma base de dados com, no mínimo, 4 colunas (variáveis):

Coluna com identificação da empresa.

Coluna com identificação do período da análise contendo, pelo menos 2 períodos, para cálculo da análise horizontal.

Coluna da variável de interesse para a análise vertical

Coluna da variável de referência para a a análise vertical.

Apresenta, como resultado, um data.frame com todas as colunas informadas, com a análise horizotal ou com a anaálise vertical.

Informações adicionais sobre como usar o pacote, orientamos acessar o menu cntdd do Blog do Projeto contabiliDados: https://contabilidados.com.br/. Ao acessar, fazer busca pelo nome da funcao ind_analiseVH

Contatos pelo email do Projeto contabiliDados: Email: contabilidados@ufersa.edu.br Siga-nos no Instagram: https://www.instagram.com/contabilidados @contabilidados

Examples

library(cntdd)

ind_analiseVH(
.dt = dt_contabil,
.contas = c("caixaEquiv", "clientesCP"),
.contaReferenciaAV = ativoTotal,
.colEmpresa = empresa,
.colPeriodo = ano,
.output = "AVAH"
)
#> # A tibble: 60 × 7
#>    empresa   ano ativoTotal AV_aplicFinanc AV_estoques AH_aplicFinanc
#>    <chr>   <int>      <int>          <dbl>       <dbl>          <dbl>
#>  1 alpha    2016     939428         0.262       0.122         NA     
#>  2 alpha    2017    1087259         0.312       0.108          0.378 
#>  3 alpha    2018    1082219         0.218       0.144         -0.307 
#>  4 alpha    2019    1463539         0.187       0.127          0.161 
#>  5 alpha    2020    2933340         0.185       0.103          0.982 
#>  6 alpha    2021    3486116         0.0680      0.134         -0.562 
#>  7 beta     2016    7995504         0.0436      0.0858        NA     
#>  8 beta     2017    8265807         0.0466      0.0915         0.104 
#>  9 beta     2018   10844797         0.0378      0.0885         0.0640
#> 10 beta     2019   13050981         0.0497      0.0835         0.583 
#> # ℹ 50 more rows
#> # ℹ 1 more variable: AH_estoques <dbl>