Capital Operacional Liquido (COL)

Essa funcao calcula o Capital Operacional Liquido de forma simplificada, pois nao calcula o ativo e passivo operacional, mas faz uma simplificacao pela soma do patrimonio liquido com a divida de curto e longo prazo, deduzidos do valor dos investimentos. Em linhas gerais, o resultado e semelhante.

ind_capitalOperLiquido(
  indicador = "Liq Seca",
  periodo = 2019:2020,
  patLiq = c(150, 200),
  investimentos = c(5, 3),
  dividasCP = c(30, 40),
  dividasLP = c(60, 90),
  atvTotal = c(500, 400),
  plot = T
)

Arguments

indicador

Um vetor tipo character com o nome do indicador

periodo

Vetor numerico indicando o periodo da analise

patLiq

Vetor com os valores do patrimonio liquido da empresa

investimentos

Vetor com os valores dos investimentos da empresa

dividasCP

Vetor com os valores de emprestimos e financiamentos de curto prazo (Passivo Circulante)

dividasLP

Vetor com os valores de emprestimos e financiamentos de longo prazo (Passivo Nao Circulante)

atvTotal

Vetor com os valores do Ativo Total

plot

Mostra grafico? (TRUE/FALSE)

Details

Apresenta como resultado uma lista com 5 itens:

Grafico se o parametro plot for TRUE ou T, mostra um grafico com a evolucao capital operacional líquido da empresa durante os periodos. Se for FALSE ou F, o grafico nao e apresentado;

Contas que corresponde ao banco de dados com as contas informadas para calculo do indicador;

Indice o capital operacional liquido dos periodos informados;

Analise Vertical Analise Vertical das contas informadas no item 1. Contas de resultado terao sua analise vertical em relacao a receita total e contas patrimoniais terao sua analise vertical em relacao ao ativo total;

Analise Horizontal Analise Horizontal das contas informadas no item 1.

Todos os itens da lista sao bancos de dados no formato tibble que podem ser usados individualmente durante o processo de analise de dados.

Examples