Indice de Liquidez Corrente

Essa funcao calcula o Indice de Liquidez Corrente baseado em vetores relativos as contas de ativo circulante e passivo circulante.

ind_liqCorrente(
  indicador = "Liq Corrente",
  periodo = 2019:2020,
  cxEquiv = c(8, 10),
  estoque = c(150, 200),
  ctaRecCP = c(400, 300),
  outAtvCirc = c(1, 3),
  fornec = c(50, 20),
  dividasCP = c(30, 40),
  outPasCirc = c(10, 8),
  atvTotal = c(900, 800),
  plot = T
)

Arguments

indicador

Um vetor tipo character com o nome do indicador

periodo

Vetor numérico indicando o periodo da analise

cxEquiv

Vetor com os valores do caixa e equivalentes de caixa (Ativo Circulante)

estoque

Vetor com os valores do estoque (Ativo Circulante)

ctaRecCP

Vetor com os valores do contas a receber (Ativo Circulante)

outAtvCirc

Vetor com os valores de outros ativos circulantes (Ativo Circulante)

fornec

Vetor com os valores de fornecedores (Passivo Circulante)

dividasCP

Vetor com os valores de empréstimos e financiamentos de curto prazo (Passivo Circulante)

outPasCirc

Vetor com os valores de outros passivos circulantes (Passivo Circulante)

atvTotal

Vetor com os valores do Ativo Total

plot

Mostra grafico? (TRUE/FALSE)

Details

Apresenta como resultado uma lista com 5 itens:

Grafico se o parametro plot for TRUE ou T, mostra um grafico com a evolucao da Liquidez corrente da empresa durante os periodos. Se for FALSE ou F, o grafico nao é apresentado;

Contas que corresponde ao banco de dados com as contas informadas para calculo do indicador;

Indice o indice de Liquidez Corrente dos periodos informados;

Analise Vertical Analise Vertical das contas informadas no item 1. Conta de resultado terao sua analise vertical em relacao a receita total e contas patrimoniais terao sua analise vertical em relacao ao ativo total;

Analise Horizontal Analise Horizontal das contas informadas no item 1.

Todos os itens da lista sao bancos de dados no formato tibble que podem ser usados individualmente durante o processo de analise de dados.

Examples