Índice de Liquidez Seca

Esse índice corresponde à capacidade de um vetor numérico

ind_liqSeca(
  indicador = "Liq Seca",
  periodo = 2019:2020,
  cxEquiv = c(8, 10),
  estoque = c(150, 200),
  ctaRecCP = c(400, 300),
  outAtvCirc = c(1, 3),
  fornec = c(50, 20),
  dividasCP = c(30, 40),
  outPasCirc = c(10, 8),
  dividasLP = c(80, 10),
  outPasNCirc = c(5, 3),
  atvTotal = c(900, 800),
  plot = T
)

Arguments

indicador

Um vetor tipo character com o nome do indicador

periodo

Vetor numérico indicando o período da análise

cxEquiv

Vetor com os valores do caixa e equivalentes de caixa (Ativo Circulante)

estoque

Vetor com os valores do estoque (Ativo Circulante)

ctaRecCP

Vetor com os valores do contas a receber (Ativo Circulante)

outAtvCirc

Vetor com os valores de outros ativos circulantes (Ativo Circulante)

fornec

Vetor com os valores de fornecedores (Passivo Circulante)

dividasCP

Vetor com os valores de empréstimos e financiamentos de curto prazo (Passivo Circulante)

outPasCirc

Vetor com os valores de outros passivos circulantes (Passivo Circulante)

atvTotal

Vetor com os valores do Ativo Total

plot

Mostra gráfico? (TRUE/FALSE)

Details

O índice de Liquidez Seca revela a capacidade de pagamento de dívidas de curto prazo, considerando que a empresa não consiga reverter o valor de seus estoques em dinheiro. Relaciona, portanto, quanto uma empresa possui de ativo (circulante), deduzido do valor de seus estoques, com o total de seu passivo circulante. É representado pela fórmula:

$$\frac{AC - Est}{PC}$$ em que: AC é o Ativo Circulante, Est é o estoque e PC é o Passivo Circulante

Para melhorar o processo de análise, os valores de Ativos Circulante e Passivo Circulante foram desmembrados, respectivamente para \(AC = cxEquiv + estoque + ctaRecCP + outAtvCirc\) e para \(PC = fornec + dividasCP + outPasCirc\). As contas são descritas acima.

Assim, tem-se que a liquidez seca corresponde a:

$$\frac{(cxEquiv + ctaRecCP + outAtvCirc)}{(fornec + dividasCP + outPasCirc)}$$

A equação não contempla o estoque!

Indicamos essa análise quando a empresa tiver algum indicativo de que terar dificuldade com giro de seus estoques. Em codições normais, espera-se que as empresas não passem dificuldade de liquidez (condição de quitar suas dívidas) por falta de conversão do estoque em caixa.

Examples

library(cntdd)
suppressMessages(suppressWarnings(library(dplyr)))
suppressMessages(suppressWarnings(library(tidyr)))
suppressMessages(suppressWarnings(library(ggplot2)))

## Usando Vetores
ind_liqSeca(indicador = "Liquidez Seca",
periodo = 2018:2020,
estoque = c(2000,3000,4000),
estoque = c(2000,3000,4000),
atvCirc = c(5000,6000,7000),
pasCirc = c(1000,2000,3000),
atvTotal = c(10000, 12000, 14000),
plot = T)
#> Error in ind_liqSeca(indicador = "Liquidez Seca", periodo = 2018:2020,     estoque = c(2000, 3000, 4000), estoque = c(2000, 3000, 4000),     atvCirc = c(5000, 6000, 7000), pasCirc = c(1000, 2000, 3000),     atvTotal = c(10000, 12000, 14000), plot = T): formal argument "estoque" matched by multiple actual arguments

## Usando um data.frame (informar valor zero quando valor não existente)

dadosAlpha <- dt_contabil %>% filter(empresa == "alpha")

ind_liqSeca(
indicador  = "Liquidez Seca",
periodo    = dadosAlpha$periodo,
cxEquiv    = dadosAlpha$cxEquiv,
estoque    = dadosAlpha$estoque,
ctaRecCP   = dadosAlpha$ctaRecCP,
outAtvCirc = dadosAlpha$outAtvCirc,
fornec     = dadosAlpha$outPasCirc,
dividasCP  = dadosAlpha$dividasCP,
outPasCirc = dadosAlpha$outPasCirc,
atvTotal   = dadosAlpha$atvTotal,
plot = F)
#> Error in data.frame(Periodo = periodo, ratio = ratio): arguments imply differing number of rows: 0, 4