Custo Medio de Capital Ponderado (WACC ou CMPC)

Essa funcao calcula o Custo Medio de Capital Ponderado baseado em vetores relativos ao custo da divida, peso da divida e o beneficio fiscal do imposto de renda, custo de capital do sócio, peso do capital do sócio.

ind_wacc(
  indicador = "WACC",
  periodo = 2019:2020,
  empFinCP = c(200, 300),
  empFinLP = c(100, 200),
  patLiq = c(400, 600),
  despJuros = c(30, 40),
  receita = c(1000, 1200),
  atvTotal = c(4000, 5000),
  ir = 0.34,
  capm = T,
  ke = 0.15,
  beta = 1.3,
  rf = 0.07,
  rm = 0.13,
  plot = T
)

Arguments

indicador

Um vetor tipo character com o nome do indicador

periodo

Vetor numerico indicando o periodo da analise

empFinCP

Vetor com os valores de emprestimos e financiamentos de curto prazo (Passivo Circulante)

empFinLP

Vetor com os valores de emprestimos e financiamentos de longo prazo (Passivo Nao Circulante)

patLiq

Vetor com os valores do Patrimônio Liquido

despJuros

Vetor com os valores das despesas com juros (DRE/Notas Explicativas)

receita

Vetor com os valores da receita liquida (Demonstracao de Resultado - DRE)

atvTotal

Vetor com os valores do Ativo Total

ir

Valor relativo a aliquota do Imposto de Renda

capm

Se TRUE, calcula ke pela fórmula rf + beta * (rm - rf). Se FALSE, calcula pelo parametro ke

ke

Valor relativo ao custo de capital do sócio

beta

Valor relativo ao beta (medida de volatilidade ou risco sistematico) de um titulo ou portifólio comparado ao mercado como um todo

rf

Valor relativo a taxa de juros livre de risco ou risco baixo

rm

Valor relativo a taxa de remuneracao paga pelo mercado

plot

Se TRUE, Mostra grafico do indicador. Se FALSE, nao mostra o grafico (TRUE/FALSE)

Details

Apresenta como resultado uma #' lista com 5 itens:

Grafico se o parametro plot for TRUE ou T, mostra um grafico com a evolucao do Custo Medio de Capital Ponderado dos sócios e banco. Se for FALSE ou F, o grafico nao e apresentado;

Contas que corresponde ao banco de dados com as contas informadas para calculo do indicador;

Indice o indice de Custo Medio de Capital Ponderado dos periodos informados;

Analise Vertical Analise Vertical das contas informadas no item 1. Conta de resultado terao sua analise vertical em relacao a receita total e contas patrimoniais terao sua analise vertical em relacao ao ativo total;

Analise Horizontal Analise Horizontal das contas informadas no item 1.

Todos os itens da lista sao bancos de dados no formato tibble que podem ser usados individualmente durante o processo de analise de dados.

Examples