Correlação de Pearson e Spearman

Gera matriz de correlação de Pearson e/ou Spearman com suas respectivas significâncias estatísticas.

stat_correl(
  dt = cntdd::dt_contabil[4:8],
  pearsonLower = T,
  ommitLower = F,
  ommitUpper = F,
  digits = 2,
  decimal = ","
)

Arguments

dt

Um data.frame

pearsonLower

Se TRUE ou T, a correlação de Pearson e disposta na diagonal inferior da Matriz e a correlação de Spearman na diagonal superior. Se FALSE OU F, a disposição das correlações é inversa. (Padrão: T)

ommitLower

Se TRUE ou T, a diagonal INFERIOR é omitida. Se FALSE OU F, e mostrada. (Padrão: F)

ommitUpper

Se TRUE ou T, a diagonal SUPERIOR é omitida. Se FALSE OU F, e mostrada. (Padrão: F)

digits

Número de dígitos da correlação (Padrão: 2)

decimal

Caracter para casa decimal (Padrão: ,)

Details

Informações adicionais sobre como usar o pacote, orientamos acessar o menu cntdd do Blog do Projeto contabiliDados: https://contabilidados.quarto.pub/. Ao acessar, fazer busca pelo nome da função stat_correl.

Contatos pelo email do Projeto contabiliDados: Email: contabilidados@ufersa.edu.br Siga-nos no Instagram: https://www.instagram.com/contabilidados @contabilidados

Examples

library(cntdd)

stat_correl(
 dt = cntdd::dt_contabil[4:8],
 pearsonLower = TRUE,
 ommitLower = FALSE,
 ommitUpper = FALSE,
 digits = 2,
 decimal = ","
 )
#>               aplicFinanc clientesCP estoques outrosAtvCirc atvCirc
#> aplicFinanc                   0,32**    0,22*        0,27**  0,28**
#> clientesCP          0,22*             0,86***       0,85*** 0,93***
#> estoques          0,34***    0,78***                0,91*** 0,97***
#> outrosAtvCirc     0,38***    0,69***  0,90***               0,93***
#> atvCirc           0,38***    0,92***  0,94***       0,87***