Teste de Médias ou Medianas

Esta função realiza o teste de diferença de médias ou mediana de dois vetores numéricos e apresenta os resultados em uma tabela.

stat_meanTest(value_Y, name_Y = "Y", value_X, name_X = "X", pvalor = 0.05)

Arguments

value_Y

Um vetor numérico representando o primeiro grupo

name_Y

Um nome para o vetor numérico que representa o primeiro grupo

value_X

Um vetor numérico representando o segundo grupo

name_X

Um nome para o vetor numérico que representa o segundo grupo

pvalor

O valor crítico a ser considerado nos testes. O padrão é 5%

Value

Um data.frame com a análise sobre o teste de média/mediana dos dois grupos de valores.

Details

A escolha entre o teste de média ou mediana é realizada de forma automática, de acordo com a distribuição das variáveis numéricas informadas.

Informações adicionais sobre como usar o pacote, orientamos acessar o menu cntdd do Blog do Projeto contabiliDados: https://contabilidados.quarto.pub/. Ao acessar, fazer busca pelo nome da função stat_meanTest.

Contatos pelo email do Projeto contabiliDados: Email: contabilidados@ufersa.edu.br Siga-nos no Instagram: https://www.instagram.com/contabilidados @contabilidados

Examples

library(cntdd)
library(dplyr)
#> 
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag
#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union
library(nortest)

dt <- na.omit(cntdd::dt_contabil)

dt <- cntdd::utl_createGroup(
 bd = dt, col_Value = "receita", col_Grp = "grp.Receita",
 n_grp = 2) %>%
 select(grp.Receita, caixaEquiv)

 grupo1 <- dt %>% filter(grp.Receita == "receita_1") %>% select(caixaEquiv) %>% pull()
 grupo2 <- dt %>% filter(grp.Receita == "receita_2") %>% select(caixaEquiv) %>% pull()

 stat_meanTest(
   value_Y = grupo1, name_Y = "Menor Receita",
   value_X = grupo2, name_X = "Maior Receita"
   )
#>                                                                                      Resultados
#> pValue Normalidade de Menor Receita                                                       0,000
#> pValue Normalidade de Maior Receita                                                       0,001
#> Resultado Normalidade                      Pelo menos um dos grupos nao tem distribuicao normal
#> Teste Usado                                                                        Mann-Whitney
#> pValue Mann-Whitney                                                                       0,000
#> Resultado Mann-Whitney                                                      Medianas diferentes
#> Media de Menor Receita                                                               141532,556
#> Media de Maior Receita                                                              1178848,346
#> Media: Menor Receita menos Maior Receita                                           -1037315,791
#> Mediana de Menor Receita                                                              41235,000
#> Mediana de Maior Receita                                                             860363,500
#> Mediana: Menor Receita menos Maior Receita                                          -819128,500
#> Qde de Obs                                                Menor Receita: 27 | Maior Receita: 26