Teste de Independência

Essa função faz o teste de independência entre variáveis categóricas e retorna uma lista com a análise correspondente.

stat_testaIndependencia(bd, correcao = T, ...)

Arguments

bd

Um data.frame com as variáveis categóricas

correcao

Se TRUE, fará a correcao dos erros

...

a lista de variáveis categóricas de interesse

Details

Informações adicionais sobre como usar o pacote, orientamos acessar o menu cntdd do Blog do Projeto contabiliDados: https://contabilidados.quarto.pub/. Ao acessar, fazer busca pelo nome da função stat_testaIndependencia

Contatos pelo email do Projeto contabiliDados: Email: contabilidados@ufersa.edu.br Siga-nos no Instagram: https://www.instagram.com/contabilidados @contabilidados

Examples

library(cntdd)
library(dplyr)

dt <-
cntdd::dt_cvmB3 %>%
transmute(
  especieControleAcionario,
  banco = if_else(setorCVM == "Bancos", "Banco", "Nao Banco")
  )

stat_testaIndependencia(bd = dt, correcao = FALSE, especieControleAcionario, banco)
#> $`Tabela de Frequencia`
#>                         banco
#> especieControleAcionario Banco Nao Banco
#>                              0         3
#>      Estatal                 9        27
#>      Estatal Holding         0         8
#>      Estrangeiro             4        26
#>      Estrangeiro Holding     4        30
#>      Privado                28       885
#>      Privado Holding         0       205
#> 
#> $`Analise Qui2`
#> [1] "Ha associacao entre os dois grupos (p-Valor: 0.00)"
#> 
#> $`Valores Esperados`
#>                         banco
#> especieControleAcionario      Banco  Nao Banco
#>                           0.1098454   2.890155
#>      Estatal              1.3181448  34.681855
#>      Estatal Holding      0.2929211   7.707079
#>      Estrangeiro          1.0984540  28.901546
#>      Estrangeiro Holding  1.2449146  32.755085
#>      Privado             33.4296176 879.570382
#>      Privado Holding      7.5061025 197.493897
#> 
#> $`Residuos Padronizados Ajustados`
#>                         banco
#> especieControleAcionario      Banco  Nao Banco
#>                          -0.3380818  0.3380818
#>      Estatal              6.9189527 -6.9189527
#>      Estatal Holding     -0.5532144  0.5532144
#>      Estrangeiro          2.8556498 -2.8556498
#>      Estrangeiro Holding  2.5512756 -2.5512756
#>      Privado             -1.8868441  1.8868441
#>      Privado Holding     -3.0579706  3.0579706
#> 
#> $`Ponto de Corte`
#> [1] 2.6901
#> 
#> $`Residuos Significativos`
#>                         banco
#> especieControleAcionario Banco Nao Banco
#>                          FALSE     FALSE
#>      Estatal              TRUE      TRUE
#>      Estatal Holding     FALSE     FALSE
#>      Estrangeiro          TRUE      TRUE
#>      Estrangeiro Holding FALSE     FALSE
#>      Privado             FALSE     FALSE
#>      Privado Holding      TRUE      TRUE
#> 
#> $`V de Cramer`
#> [1] 0.2385
#>