Máscara CNPJ

Gera número CNPJ no formato desejado.

utl_CNPJ_Mascara(cnpj = "21350241000322", completo = T)

Arguments

cnpj

string com número do CNPJ

completo

Se TRUE ou T, o output mostrar o CNPJ no formato 99.999.999/0001-99. Se FALSE ou F, o output será no formato 99999999000199.

Details

Essa função formata string com numero do CNPJ que possua até 14 caracteres com o número do CNPJ completo (99.999.999/0001-99) ou simplificado (999999990000199).

Informações adicionais sobre como usar o pacote, orientamos acessar o menu cntdd do Blog do Projeto contabiliDados: https://contabilidados.quarto.pub/. Ao acessar, fazer busca pelo nome da função utl_CNPJ_Mascara

Contatos pelo email do Projeto contabiliDados: Email: contabilidados@ufersa.edu.br Siga-nos no Instagram: https://www.instagram.com/contabilidados @contabilidados

Examples

library(cntdd)

utl_CNPJ_Mascara(cnpj = "21350241000322", completo = TRUE)
#>       21350241000322 
#> "21.350.241/0003-22"