Ajuste de Valores pela Inflação

Faz o ajuste do valor de uma conta pelo valor do índice de inflação informado.

utl_ajustaContaInflacao(
  periodo = 2019:2023,
  vrConta = c(100, 110, 120, 130, 140),
  indiceInflacao = c(4.31, 4.52, 10.06, 5.79, 4.62),
  periodoBase = 2023,
  apenasVetor = T
)

Arguments

periodo

Vetor contendo o período dos dados

vrConta

Vetor contendo o valor da conta que pretende atualizar pela inflação

indiceInflacao

Valor percentual do indice de inflacao do período. Se o valor for 2,5%, por exemplo, informar 2.5

periodoBase

Período específico para o qual deseja atualizar os valores

apenasVetor

Se TRUE ou T (padrão), o resultado é um vetor com os valores atualizados. Se FALSE ou F, o resultado é um data.frame com todos os valores envolvidos no cálculo do ajuste à inflação

Details

Informações adicionais sobre como usar o pacote, orientamos acessar o menu cntdd do Blog do Projeto contabiliDados: https://contabilidados.quarto.pub/. Ao acessar, fazer busca pelo nome da função utl_ajustaContaInflacao.

Contatos pelo email do Projeto contabiliDados: Email: contabilidados@ufersa.edu.br Siga-nos no Instagram: https://www.instagram.com/contabilidados @contabilidados

Examples

library(cntdd)

utl_ajustaContaInflacao(
periodo = 2019:2023,
vrConta = c(100, 110, 120, 130, 140),
indiceInflacao = c(4.31, 4.52, 10.06, 5.79, 4.62),
periodoBase = 2023, apenasVetor = TRUE)
#> Aviso: Certifique-se de que os dados de inflacao estao em percentual.
#>             Exemplo: Se for 2,5%, informar o valor 2.5 no argumento 'indiceInflacao'
#> [1] 100.2972 110.1052 114.0687 128.5622 140.0000