Baixa dados Banco Central

Baixa dados da API do Banco Central

utl_baixaBCB(codSerie = 433, inicio = "01/01/2003", fim = "31/12/2017")

Arguments

codSerie

Valor do código da série no Banco Central

inicio

String contendo a data de início da série no formato dd/mm/aaaa

fim

String contendo a data final da série no formato dd/mm/aaaa

Details

Informações adicionais sobre como usar o pacote, orientamos acessar o menu cntdd do Blog do Projeto contabiliDados: https://contabilidados.quarto.pub/. Ao acessar, fazer busca pelo nome da função utl_baixaBCB.

Contatos pelo email do Projeto contabiliDados: Email: contabilidados@ufersa.edu.br Siga-nos no Instagram: https://www.instagram.com/contabilidados @contabilidados

Examples

library(cntdd)

utl_baixaBCB(codSerie = 433, inicio = "01/01/2003", fim = "31/12/2017")
#>           data valor
#> 1   01/01/2003  2.25
#> 2   01/02/2003  1.57
#> 3   01/03/2003  1.23
#> 4   01/04/2003  0.97
#> 5   01/05/2003  0.61
#> 6   01/06/2003 -0.15
#> 7   01/07/2003  0.20
#> 8   01/08/2003  0.34
#> 9   01/09/2003  0.78
#> 10  01/10/2003  0.29
#> 11  01/11/2003  0.34
#> 12  01/12/2003  0.52
#> 13  01/01/2004  0.76
#> 14  01/02/2004  0.61
#> 15  01/03/2004  0.47
#> 16  01/04/2004  0.37
#> 17  01/05/2004  0.51
#> 18  01/06/2004  0.71
#> 19  01/07/2004  0.91
#> 20  01/08/2004  0.69
#> 21  01/09/2004  0.33
#> 22  01/10/2004  0.44
#> 23  01/11/2004  0.69
#> 24  01/12/2004  0.86
#> 25  01/01/2005  0.58
#> 26  01/02/2005  0.59
#> 27  01/03/2005  0.61
#> 28  01/04/2005  0.87
#> 29  01/05/2005  0.49
#> 30  01/06/2005 -0.02
#> 31  01/07/2005  0.25
#> 32  01/08/2005  0.17
#> 33  01/09/2005  0.35
#> 34  01/10/2005  0.75
#> 35  01/11/2005  0.55
#> 36  01/12/2005  0.36
#> 37  01/01/2006  0.59
#> 38  01/02/2006  0.41
#> 39  01/03/2006  0.43
#> 40  01/04/2006  0.21
#> 41  01/05/2006  0.10
#> 42  01/06/2006 -0.21
#> 43  01/07/2006  0.19
#> 44  01/08/2006  0.05
#> 45  01/09/2006  0.21
#> 46  01/10/2006  0.33
#> 47  01/11/2006  0.31
#> 48  01/12/2006  0.48
#> 49  01/01/2007  0.44
#> 50  01/02/2007  0.44
#> 51  01/03/2007  0.37
#> 52  01/04/2007  0.25
#> 53  01/05/2007  0.28
#> 54  01/06/2007  0.28
#> 55  01/07/2007  0.24
#> 56  01/08/2007  0.47
#> 57  01/09/2007  0.18
#> 58  01/10/2007  0.30
#> 59  01/11/2007  0.38
#> 60  01/12/2007  0.74
#> 61  01/01/2008  0.54
#> 62  01/02/2008  0.49
#> 63  01/03/2008  0.48
#> 64  01/04/2008  0.55
#> 65  01/05/2008  0.79
#> 66  01/06/2008  0.74
#> 67  01/07/2008  0.53
#> 68  01/08/2008  0.28
#> 69  01/09/2008  0.26
#> 70  01/10/2008  0.45
#> 71  01/11/2008  0.36
#> 72  01/12/2008  0.28
#> 73  01/01/2009  0.48
#> 74  01/02/2009  0.55
#> 75  01/03/2009  0.20
#> 76  01/04/2009  0.48
#> 77  01/05/2009  0.47
#> 78  01/06/2009  0.36
#> 79  01/07/2009  0.24
#> 80  01/08/2009  0.15
#> 81  01/09/2009  0.24
#> 82  01/10/2009  0.28
#> 83  01/11/2009  0.41
#> 84  01/12/2009  0.37
#> 85  01/01/2010  0.75
#> 86  01/02/2010  0.78
#> 87  01/03/2010  0.52
#> 88  01/04/2010  0.57
#> 89  01/05/2010  0.43
#> 90  01/06/2010  0.00
#> 91  01/07/2010  0.01
#> 92  01/08/2010  0.04
#> 93  01/09/2010  0.45
#> 94  01/10/2010  0.75
#> 95  01/11/2010  0.83
#> 96  01/12/2010  0.63
#> 97  01/01/2011  0.83
#> 98  01/02/2011  0.80
#> 99  01/03/2011  0.79
#> 100 01/04/2011  0.77
#> 101 01/05/2011  0.47
#> 102 01/06/2011  0.15
#> 103 01/07/2011  0.16
#> 104 01/08/2011  0.37
#> 105 01/09/2011  0.53
#> 106 01/10/2011  0.43
#> 107 01/11/2011  0.52
#> 108 01/12/2011  0.50
#> 109 01/01/2012  0.56
#> 110 01/02/2012  0.45
#> 111 01/03/2012  0.21
#> 112 01/04/2012  0.64
#> 113 01/05/2012  0.36
#> 114 01/06/2012  0.08
#> 115 01/07/2012  0.43
#> 116 01/08/2012  0.41
#> 117 01/09/2012  0.57
#> 118 01/10/2012  0.59
#> 119 01/11/2012  0.60
#> 120 01/12/2012  0.79
#> 121 01/01/2013  0.86
#> 122 01/02/2013  0.60
#> 123 01/03/2013  0.47
#> 124 01/04/2013  0.55
#> 125 01/05/2013  0.37
#> 126 01/06/2013  0.26
#> 127 01/07/2013  0.03
#> 128 01/08/2013  0.24
#> 129 01/09/2013  0.35
#> 130 01/10/2013  0.57
#> 131 01/11/2013  0.54
#> 132 01/12/2013  0.92
#> 133 01/01/2014  0.55
#> 134 01/02/2014  0.69
#> 135 01/03/2014  0.92
#> 136 01/04/2014  0.67
#> 137 01/05/2014  0.46
#> 138 01/06/2014  0.40
#> 139 01/07/2014  0.01
#> 140 01/08/2014  0.25
#> 141 01/09/2014  0.57
#> 142 01/10/2014  0.42
#> 143 01/11/2014  0.51
#> 144 01/12/2014  0.78
#> 145 01/01/2015  1.24
#> 146 01/02/2015  1.22
#> 147 01/03/2015  1.32
#> 148 01/04/2015  0.71
#> 149 01/05/2015  0.74
#> 150 01/06/2015  0.79
#> 151 01/07/2015  0.62
#> 152 01/08/2015  0.22
#> 153 01/09/2015  0.54
#> 154 01/10/2015  0.82
#> 155 01/11/2015  1.01
#> 156 01/12/2015  0.96
#> 157 01/01/2016  1.27
#> 158 01/02/2016  0.90
#> 159 01/03/2016  0.43
#> 160 01/04/2016  0.61
#> 161 01/05/2016  0.78
#> 162 01/06/2016  0.35
#> 163 01/07/2016  0.52
#> 164 01/08/2016  0.44
#> 165 01/09/2016  0.08
#> 166 01/10/2016  0.26
#> 167 01/11/2016  0.18
#> 168 01/12/2016  0.30
#> 169 01/01/2017  0.38
#> 170 01/02/2017  0.33
#> 171 01/03/2017  0.25
#> 172 01/04/2017  0.14
#> 173 01/05/2017  0.31
#> 174 01/06/2017 -0.23
#> 175 01/07/2017  0.24
#> 176 01/08/2017  0.19
#> 177 01/09/2017  0.16
#> 178 01/10/2017  0.42
#> 179 01/11/2017  0.28
#> 180 01/12/2017  0.44