Cria grupos

Cria grupos baseados nos percentis de uma variavel numérica.

utl_createGroup(bd, col_Value, col_Grp, n_grp = 2)

Arguments

bd

Um data.frame

col_Value

String informando o nome da coluna referente a variável numérica sobre a qual se pretende criar os grupos

col_Grp

String com o nome da nova coluna a ser criada com os grupos

n_grp

Número de grupos, se pretende criar grupos com o mesmo numero de observações ou um vetor contendo os percentis para cada grupo. Padrão igual 2.

Details

Útil para tornar variáveis numéricas em categóricas para realização de testes de análise exploratória.

Informações adicionais sobre como usar o pacote, orientamos acessar o menu cntdd do Blog do Projeto contabiliDados: https://contabilidados.quarto.pub/. Ao acessar, fazer busca pelo nome da função utl_createGroup.

Contatos pelo email do Projeto contabiliDados: Email: contabilidados@ufersa.edu.br Siga-nos no Instagram: https://www.instagram.com/contabilidados @contabilidados

See also

Examples

library(cntdd)
library(dplyr)

dt <- cntdd::dt_contabil %>% filter(!is.na(ativoTotal))

utl_createGroup(bd = dt, col_Value = "ativoTotal", col_Grp = "GrpAtivo", n_grp = 3) %>%
select(empresa, ano, ativoTotal, GrpAtivo) %>% head()
#>   empresa  ano ativoTotal     GrpAtivo
#> 1   alpha 2016     939428 ativoTotal_1
#> 2   alpha 2017    1087259 ativoTotal_1
#> 3   alpha 2018    1082219 ativoTotal_1
#> 4   alpha 2019    1463539 ativoTotal_1
#> 5   alpha 2020    2933340 ativoTotal_1
#> 6   alpha 2021    3486116 ativoTotal_2