Limpeza de CNPJs em um DataFrame

Esta funcão recebe um DataFrame e um nome de coluna, e retorna uma lista de CNPJs limpos, com a remoção de caracteres especiais como pontos, barras e hífens.

utl_limpaCNPJ(df, nome_coluna)

Arguments

df

DataFrame que contém a coluna de CNPJs.

nome_coluna

O nome da coluna no DataFrame que contém os CNPJs.

Value

Um vetor de CNPJs limpos.

Details

A máscara do CNPJ deve esta no formato 99.999.999/9999-99.

Informações adicionais sobre como usar o pacote, orientamos acessar o menu cntdd do Blog do Projeto contabiliDados: https://contabilidados.quarto.pub/. Ao acessar, fazer busca pelo nome da função utl_limpaCNPJ

Contatos pelo email do Projeto contabiliDados: Email: contabilidados@ufersa.edu.br Siga-nos no Instagram: https://www.instagram.com/contabilidados @contabilidados

Examples

library(cntdd)

utl_limpaCNPJ(cntdd::dt_cvmB3, cnpj)
#>    [1] "00000000000191" "15124464000187" "05707616000110" "15144017000190"
#>    [5] "61856571000117" "17245234000100" "60730348000166" "21255567000189"
#>    [9] "82981671000145" "33433665000148" "88610191000154" "92690783000109"
#>   [13] "33009911000139" "60444437000146" "19527639000158" "61079117000105"
#>   [17] "97191902000194" "33050071000158" "82981929000103" "60851615000153"
#>   [21] "21254073000180" "61067161000197" "33050196000188" "60500139000126"
#>   [25] "20933354000104" "92665611000177" "61351532000168" "84208123000102"
#>   [29] "61156931000178" "82643537000134" "05721582000119" "82982075000180"
#>   [33] "92702067000196" "08884213000135" "61065298000102" "61584140000149"
#>   [37] "33014556000196" "92660570000126" "84683671000194" "95426862000197"
#>   [41] "88611983000143" "17159005000164" "82640558000104" "92929520000100"
#>   [45] "01838723000127" "58128174000114" "61585865000151" "82641325000118"
#>   [49] "82636986000155" "60637238000154" "60637238000154" "91983056000169"
#>   [53] "19526748000150" "92749225000163" "60605664000106" "60665981000118"
#>   [57] "30540991000166" "43948405000169" "61082004000150" "28127603000178"
#>   [61] "84683374000149" "07450604000189" "56720428000163" "61101895000145"
#>   [65] "33376989000191" "33467572000134" "92781335000102" "61092037000181"
#>   [69] "04912242000102" "17184037000110" "61156113000175" "33042730000104"
#>   [73] "92791243000103" "60643228000121" "61409892000173" "86375425000109"
#>   [77] "60882628000190" "04902979000144" "60746948000112" "33592510000154"
#>   [81] "33066408000115" "33200049000147" "51466860000156" "60872504000123"
#>   [85] "17157777000167" "60664810000174" "61080313000191" "61084984000120"
#>   [89] "33147315000115" "61088894000108" "85778074000106" "33412081000196"
#>   [93] "60543816000193" "60543816000193" "07857093000114" "33041260065290"
#>   [97] "33041260065290" "33041260065290" "33228024000151" "33102476000192"
#>  [101] "50564053000103" "33541368000116" "56994924000105" "56992423000190"
#>  [105] "46014635000149" "05711813000103" "84684455000163" "33039223000111"
#>  [109] "88611835000129" "02183757000193" "60476884000187" "23041875000137"
#>  [113] "92012467000170" "33035536000100" "82640632000184" "07206816000115"
#>  [117] "61410395000195" "56643018000166" "80227184000166" "17193806000146"
#>  [121] "61374161000130" "17155730000164" "50926955000142" "86047198000184"
#>  [125] "82981812000120" "83296889000123" "33000167000101" "92721232000157"
#>  [129] "33386210000119" "33256439000139" "17162082000173" "17167412000113"
#>  [133] "71208516000174" "88610126000129" "60651809000105" "60894730000105"
#>  [137] "07237373000120" "91333666000117" "76483817000120" "84683762000120"
#>  [141] "60933603000178" "83878892000155" "01543032000104" "89723993000133"
#>  [145] "03467321000199" "60884319000159" "08402943000152" "33000118000179"
#>  [149] "90724345000180" "17159229000176" "33113309000147" "59105999000186"
#>  [153] "33111246000190" "47960950000121" "06272793000184" "28500320000120"
#>  [157] "08262343000136" "61189288000189" "13017462000163" "78588415000115"
#>  [161] "17167396000169" "15139629000194" "04913711000108" "47080619000117"
#>  [165] "45987245000192" "91495499000100" "60889128000180" "15141799000103"
#>  [169] "52736840000110" "60398369000126" "60398369000479" "45453214000151"
#>  [173] "15151269000146" "84429695000111" "33453598000123" "17161241000115"
#>  [177] "33611500000119" "08324196000181" "97837181000147" "15126626000116"
#>  [181] "00001180000126" "61190096000192" "91669747000192" "91669747000192"
#>  [185] "86550951000150" "04895728000180" "76484013000145" "23236821000127"
#>  [189] "13009717000146" "76500180000132" "84693183000168" "17281106000103"
#>  [193] "76535764000143" "92715812000131" "09095183000140" "01548981000179"
#>  [197] "84429752000162" "88613658000110" "60631090000140" "60840055000131"
#>  [201] "59104513000195" "00000208000100" "61186680000174" "17314329000120"
#>  [205] "10835932000108" "92754738000162" "25329319000196" "81304727000164"
#>  [209] "92660604000182" "92802784000190" "15115504000124" "61565222000146"
#>  [213] "02800026000140" "07203904000163" "61532644000115" "60770336000165"
#>  [217] "61231155000123" "61486650000183" "60394723000144" "50746577000115"
#>  [221] "15101405000193" "84683408000103" "17162280000137" "51468056000106"
#>  [225] "33161340000153" "04909479000134" "59285411000113" "61022042000118"
#>  [229] "07253321000147" "09123654000187" "61024352000171" "61784278000191"
#>  [233] "52123916000132" "34020354000110" "92825900000196" "76255926000190"
#>  [237] "92693019000189" "96418264021802" "01616929000102" "07275159000168"
#>  [241] "22677520000176" "62002886000160" "16233389000155" "62063177000194"
#>  [245] "62070362000106" "76627504000106" "28152650000171" "33858168000191"
#>  [249] "62136254000199" "62232889000190" "32286411000181" "62307848000115"
#>  [253] "62258884000136" "22266175000188" "17359233000188" "00070698000111"
#>  [257] "07332190000193" "04200572000175" "16234171000115" "33958695000178"
#>  [261] "08424178000171" "33938119000169" "92934215000106" "13788120000147"
#>  [265] "52548435000179" "71673990000177" "62984091000102" "04986188000140"
#>  [269] "33130691000105" "89850341000160" "09116278000101" "34265561000134"
#>  [273] "16234627000147" "45365558000109" "82508433000117" "07047251000170"
#>  [277] "07040108000157" "13608583000180" "43185362000107" "61695227000193"
#>  [281] "34274233000102" "34274233000102" "87870952000144" "87043832000173"
#>  [285] "42150391000170" "43283811000150" "43358647000100" "43425008000102"
#>  [289] "16590234000176" "25869736000121" "00336701000104" "33040601000187"
#>  [293] "13552070000102" "34169557000172" "75609123000123" "42278473000103"
#>  [297] "42278473000103" "42278291000124" "43776517000180" "16670085000155"
#>  [301] "16676520000159" "87456562000122" "82901000000127" "42331462000131"
#>  [305] "42331462000131" "46119855000137" "00383281000109" "07281413000130"
#>  [309] "33352394000104" "47193149000106" "19296342000129" "14308514000113"
#>  [313] "19403252000190" "19378769000176" "19443985000158" "14388334000199"
#>  [317] "89096457000155" "61192373000104" "49263189000102" "05441787000140"
#>  [321] "83475913000191" "29978814000187" "29980141000108" "89637490000145"
#>  [325] "89547269000104" "15413826000150" "30306294000145" "51218147000193"
#>  [329] "14807945000124" "51128999000190" "59981829000165" "43470988000165"
#>  [333] "30723886000162" "89463822000112" "75315333000109" "44689123000157"
#>  [337] "02635522000195" "27093558000115" "45242914000105" "16614075000100"
#>  [341] "06626253000151" "87762563000103" "47509120000182" "47508411000156"
#>  [345] "49669856000143" "49669856000143" "15527906000136" "51928174000150"
#>  [349] "70216429000104" "53113791000122" "07334600000135" "28195667000106"
#>  [353] "33937681000330" "33937681000178" "53518684000184" "90076886000140"
#>  [357] "53790218000153" "00080671000100" "29780061000109" "00994533000136"
#>  [361] "14388520000128" "07636657000199" "29388352000148" "54526082000131"
#>  [365] "61065751000180" "28573830000127" "01468594000122" "90400888000142"
#>  [369] "78876950000171" "90441460000148" "29950060000157" "79313086000162"
#>  [373] "22164990000136" "79430682000122" "22261473000185" "79655916000130"
#>  [377] "14998371000119" "56444250000175" "56806656000150" "58119199000151"
#>  [381] "12049631000184" "22770366000182" "16404287000155" "58430828000160"
#>  [385] "91820068000172" "31895683000116" "80680093000181" "58877812000108"
#>  [389] "32110314000133" "59418806000147" "24962466000136" "24956666000186"
#>  [393] "59717553000102" "59789545000171" "81106957000119" "60537263000166"
#>  [397] "25089509000183" "81243735000148" "92859974000143" "32681371000172"
#>  [401] "35791391000194" "93167658000173" "81616807000155" "62092739000128"
#>  [405] "39115514000128" "26386169000115" "82120676000183" "65654303000173"
#>  [409] "82373077000171" "64904295000103" "05721735000128" "93828986000173"
#>  [413] "36461242000120" "67203208000189" "67010660000124" "67437434000124"
#>  [417] "67571414000141" "67620377000114" "67915785000101" "48344725000719"
#>  [421] "42771949000135" "68400225000179" "94789468000150" "89086144000116"
#>  [425] "69720910000145" "84583137000106" "36542025000164" "96298013000168"
#>  [429] "84496066000104" "73903718000161" "71476527000135" "73178600000118"
#>  [433] "71550388000142" "74409467000126" "00095147000102" "00027742000100"
#>  [437] "00108786000165" "00115890000187" "00185475000108" "00192769000159"
#>  [441] "00249786000185" "00194724000113" "00416968000101" "00272185000193"
#>  [445] "00242184000104" "00257846000101" "00359742000108" "00389481000179"
#>  [449] "00514998000142" "00545378000170" "00592603000120" "00622416000141"
#>  [453] "00354053000100" "00649881000176" "00644907000193" "00743065000127"
#>  [457] "02560878000107" "00864214000106" "00924432000199" "00924429000175"
#>  [461] "00866577000180" "00861626000192" "00873894000124" "00897864000158"
#>  [465] "00938574000105" "00974211000125" "01027058000191" "92692979000124"
#>  [469] "00979969000156" "01104937000170" "01083200000118" "85041333000111"
#>  [473] "01107327000120" "01258943000181" "01258945000170" "01258944000126"
#>  [477] "01319044000141" "01287588000179" "01109184000195" "01355994000121"
#>  [481] "01547749000116" "01414107000149" "01417222000177" "01425787000104"
#>  [485] "01858774000110" "02724983000134" "01545826000107" "01599101000193"
#>  [489] "01386256000141" "01637895000132" "01654604000114" "01685903000116"
#>  [493] "01851771000155" "01772313000120" "01896779000138" "01896779000138"
#>  [497] "01608571000176" "01832635000118" "01938783000111" "62144175000120"
#>  [501] "01968293000168" "01971614000183" "01957772000189" "02041460000193"
#>  [505] "01989044000159" "01917705000130" "01917818000136" "02031413000169"
#>  [509] "02062747000108" "02016439000138" "02016440000162" "02107946000187"
#>  [513] "02105040000123" "02117801000167" "02084220000176" "02162616000194"
#>  [517] "02201787000185" "02149205000169" "02390206000109" "02387241000160"
#>  [521] "02193750000152" "02193750000152" "02217325000156" "02217319000107"
#>  [525] "02217319000107" "02222736000130" "02225714000123" "02351877000152"
#>  [529] "02258422000197" "02221531000130" "02286509000178" "02255187000108"
#>  [533] "02318346000168" "02338534000158" "02312699000150" "02281836000137"
#>  [537] "02291077000193" "02312666000100" "02286479000108" "02302100000106"
#>  [541] "02302101000142" "02357251000153" "02351144000118" "02328280000197"
#>  [545] "02335514000123" "02365069000144" "02365069000144" "02373517000151"
#>  [549] "02474103000119" "02382073000110" "02363918000120" "02366097000186"
#>  [553] "02415408000150" "02451848000162" "02312604000107" "02363917000186"
#>  [557] "02476710000118" "02421421000111" "02465783000104" "02565001000109"
#>  [561] "02429144000193" "02502844000166" "02535295000126" "02509491000126"
#>  [565] "02541982000154" "02509186000134" "02558074000173" "02558134000158"
#>  [569] "02558154000129" "02558124000112" "02558115000121" "02558157000162"
#>  [573] "02736470000143" "02555926000179" "02591787000139" "02600854000134"
#>  [577] "02682287000102" "02604997000114" "02604860000160" "02679185000138"
#>  [581] "02762113000150" "02762121000104" "02796775000140" "02762115000149"
#>  [585] "02762124000130" "02773542000122" "02783425000140" "02808708000107"
#>  [589] "02808481000191" "02783423000150" "02762122000140" "02846056000197"
#>  [593] "02871007000104" "03468802000119" "02664042000152" "02950811000189"
#>  [597] "02860694000162" "02992449000109" "02919555000167" "02916265000160"
#>  [601] "03014553000191" "02983428000127" "02397080000196" "03303999000136"
#>  [605] "02998609000127" "02998611000104" "02998301000181" "03025305000146"
#>  [609] "03076274000152" "03052673000183" "02643896000152" "03207703000183"
#>  [613] "03279285000130" "03220438000173" "03342704000130" "03378521000175"
#>  [617] "03410855000189" "03392498000173" "03587889000143" "00187032000148"
#>  [621] "03559006000191" "03616277000131" "03538833000107" "03601314000138"
#>  [625] "03634795000188" "03634777000104" "03758318000124" "03820456000196"
#>  [629] "03847461000192" "03767538000114" "03802115000198" "03795050000109"
#>  [633] "03818379000130" "03853896000140" "03870455000580" "03953509000147"
#>  [637] "03983431000103" "04030087000109" "04034792000176" "04030182000102"
#>  [641] "04031213000131" "04032433000180" "04038763000182" "04065791000199"
#>  [645] "04031960000170" "04069709000102" "04065053000141" "04089570000150"
#>  [649] "04200649000107" "04128563000110" "04149454000180" "08070508000178"
#>  [653] "04172213000151" "04212396000191" "04233319000118" "04310392000146"
#>  [657] "04370082000117" "04370450000127" "04368898000106" "04370282000170"
#>  [661] "04437534000130" "04368865000166" "04423567000121" "04634250000134"
#>  [665] "05197443000138" "04781143000139" "04752991000110" "04705090000177"
#>  [669] "04800265000125" "04739720000124" "04665504000181" "02932074000191"
#>  [673] "39346861035038" "04821041000108" "04668779000179" "04652792000130"
#>  [677] "14876090000193" "04992714000184" "04986320000113" "04972092000122"
#>  [681] "05341549000163" "05303439000107" "05330331000103" "05336882000184"
#>  [685] "05347060000107" "05730375000120" "05495546000184" "05889284000132"
#>  [689] "05878397000132" "05878397000132" "05878397000132" "05799312000120"
#>  [693] "04987555000120" "05918309000189" "06016348000153" "05917486000140"
#>  [697] "06018000000103" "07857850000150" "06027566000193" "11423623000193"
#>  [701] "06033448000198" "06047087000139" "06082980000103" "06057223000171"
#>  [705] "06948969000175" "06137677000152" "06164253000187" "06977745000191"
#>  [709] "06977751000149" "06347409000165" "07033882000130" "06981180000116"
#>  [713] "06981381000113" "06981176000158" "07119838000148" "07021459000110"
#>  [717] "07043628000113" "07088416000152" "07073027000153" "07256507000150"
#>  [721] "07112325000105" "07110438000171" "06349242000171" "07282377000120"
#>  [725] "07401436000131" "07437016000105" "07415333000120" "14212464000176"
#>  [729] "07710591000130" "07526557000100" "07587384000130" "07594978000178"
#>  [733] "07620094000140" "07689002000189" "07882930000165" "07628528000159"
#>  [737] "07704246000193" "07659538000151" "07671092000180" "07394616000133"
#>  [741] "07700557000184" "07718269000157" "07802794000156" "07859510000168"
#>  [745] "07809905000156" "07816890000153" "06977739000134" "07825881000129"
#>  [749] "07820907000146" "07820907000146" "07875234000121" "07859971000130"
#>  [753] "07823262000103" "07976147000160" "07949970000187" "07957093000196"
#>  [757] "07950674000104" "08170849000115" "08065557000112" "08078847000109"
#>  [761] "08343492000120" "08140464000105" "08253728000137" "08294224000165"
#>  [765] "08159965000133" "08219477000174" "08213823000107" "08312229000173"
#>  [769] "08397078000101" "08322396000103" "08281683000104" "08262121000113"
#>  [773] "08574411000100" "08364948000138" "08439659000150" "08467115000100"
#>  [777] "08493364000162" "08560444000193" "08572741000159" "08638269000100"
#>  [781] "08684547000165" "00776574000156" "00776574000156" "08534605000174"
#>  [785] "08648295000119" "08653753000108" "08613550000198" "08613550000198"
#>  [789] "08665255000185" "08773191000136" "08741499000108" "08769451000108"
#>  [793] "08764621000153" "08723106000125" "08773135000100" "08827501000158"
#>  [797] "08807432000110" "08795211000170" "08801621000186" "08801621000186"
#>  [801] "08902291000115" "08797760000183" "08811643000127" "08887330000152"
#>  [805] "08810422000134" "08835366000192" "08873873000110" "08903116000142"
#>  [809] "08822767000108" "08904090000157" "08935054000150" "09140181000126"
#>  [813] "09191251000175" "08936792000112" "08904552000136" "08937002000113"
#>  [817] "08943312000140" "08926302000105" "09054385000144" "09041168000110"
#>  [821] "07276103000128" "09074183000164" "09053134000145" "09114805000130"
#>  [825] "09083175000184" "09267996000170" "09042817000105" "09112685000132"
#>  [829] "09149503000106" "09146451000106" "09288252000132" "09295063000197"
#>  [833] "09337210000144" "09229201000130" "09257877000137" "09346601000125"
#>  [837] "09304427000158" "09313969000197" "09324949000111" "09325109000173"
#>  [841] "09326342000170" "09336431000106" "09462852000175" "09462850000186"
#>  [845] "09400937000129" "09462890000128" "09462140000156" "09305994000129"
#>  [849] "09439783000189" "10635691000153" "09359927000197" "09391823000160"
#>  [853] "09347916000197" "09358105000191" "09414761000164" "09538973000153"
#>  [857] "09478309000166" "09527023000123" "09584854000137" "09601322000160"
#>  [861] "09625321000156" "09594796000122" "09611768000176" "09611768000176"
#>  [865] "10139870000108" "09813204000116" "10234027000100" "10240186000100"
#>  [869] "10265949000177" "10215988000160" "10242478000181" "10319900000150"
#>  [873] "10285590000108" "10345009000198" "10338320000100" "10302690000197"
#>  [877] "10337440000192" "10324624000118" "10382341000122" "10464223000163"
#>  [881] "10629105000168" "10502676000137" "10440482000154" "10488244000119"
#>  [885] "10452416000102" "10608405000160" "10531501000158" "10664513000150"
#>  [889] "10571175000102" "10603674000134" "10661519000174" "10647979000148"
#>  [893] "10670314000155" "10678505000163" "10753164000143" "10739356000103"
#>  [897] "10760260000119" "10846981000146" "10847761000137" "10747021000129"
#>  [901] "10807374000177" "10894123000177" "10826798000189" "10851805000100"
#>  [905] "10841050000155" "11197023000154" "11102250000159" "11040082000114"
#>  [909] "11094546000175" "40263170000183" "11257352000143" "11193684000101"
#>  [913] "11181400000167" "11284204000118" "11274829000107" "11396633000187"
#>  [917] "11306912000102" "11421994000136" "11395624000171" "12104241000402"
#>  [921] "11818783000131" "11716471000117" "11566501000156" "11992680000193"
#>  [925] "11858621000127" "11616840000108" "11995745000154" "11995753000109"
#>  [929] "11725176000127" "11669021000110" "11721921000160" "15073274000188"
#>  [933] "12395063000173" "12646054000108" "12126500000153" "12139922000163"
#>  [937] "12181987000177" "12078880000106" "12394916000152" "12395010000152"
#>  [941] "12610097000133" "12130744000100" "12261588000116" "12394972000197"
#>  [945] "13720918000157" "12610165000164" "12394936000123" "12091809000155"
#>  [949] "12498904000178" "12499618000127" "12499629000107" "12394991000113"
#>  [953] "13720879000198" "12610116000121" "12108897000150" "12320349000190"
#>  [957] "12610072000130" "12492734000114" "12300288000107" "12420164000157"
#>  [961] "12489315000123" "12648266000124" "12528708000107" "12610764000188"
#>  [965] "12648327000153" "12668780000121" "12668792000156" "12592795000153"
#>  [969] "12592746000110" "12592831000189" "12919786000124" "12827269000125"
#>  [973] "12979898000170" "13324184000197" "13349677000181" "13457942000145"
#>  [977] "13217485000111" "13217485000111" "13407617000178" "13407573000186"
#>  [981] "13444994000187" "13522948000159" "14333683000103" "13878847000115"
#>  [985] "13584310000142" "13483537000100" "13574594000196" "13574594000196"
#>  [989] "13776362000110" "14191427000129" "14200537000100" "14200586000142"
#>  [993] "14200618000100" "14200699000148" "14042551000122" "14969641000163"
#>  [997] "14882295000181" "14127813000151" "94813102000170" "13783221000478"
#> [1001] "14110585000107" "14206371000139" "19688956000156" "14292540000109"
#> [1005] "15114494000102" "14785152000151" "13270520000166" "15229110000105"
#> [1009] "15090690000194" "14962356000110" "15011696000129" "15463423000115"
#> [1013] "15484093000144" "16103773000133" "15494541000190" "15483430000189"
#> [1017] "15377572000161" "15688041000190" "15440708000130" "17346997000139"
#> [1021] "15578569000106" "16811931000100" "15914972000169" "15675715000111"
#> [1025] "15912764000120" "16922038000151" "16838421000126" "17234244000131"
#> [1029] "14776142000150" "17344597000194" "17568683000180" "17632564000140"
#> [1033] "18572225000188" "17943792000130" "03494776000101" "18174270000184"
#> [1037] "18328118000109" "18593815000197" "18483666000103" "18494485000182"
#> [1041] "18903785000178" "19133012000112" "19014221000147" "19276528000116"
#> [1045] "19208022000170" "19527586000175" "19288959000100" "09347516000181"
#> [1049] "19521322000104" "19511433000130" "19642306000170" "19851496000135"
#> [1053] "20445827000116" "19726048000100" "19853511000184" "19796586000170"
#> [1057] "20451953000183" "20247322000147" "20258278000170" "20646442000117"
#> [1061] "20512706000140" "20514555000169" "20514590000188" "20553221000102"
#> [1065] "21110778000123" "21331849000118" "21314559000166" "21262638000170"
#> [1069] "21240146000184" "21414457000112" "21610416000100" "21389501000181"
#> [1073] "21903297000175" "21987041000193" "21987053000118" "21987068000186"
#> [1077] "21987030000103" "21728500000114" "21728083000100" "22131736000131"
#> [1081] "21976484000189" "21948105000147" "21976510000179" "22163297000149"
#> [1085] "22495265000140" "22543331000100" "23096269000119" "22983283000171"
#> [1089] "23043212000151" "22902694000195" "23373000000132" "23438929000100"
#> [1093] "23864409000151" "24170211000130" "24230275000180" "24396489000120"
#> [1097] "24241659000106" "24410913000144" "24410913000144" "24955657000170"
#> [1101] "24717074000101" "24624490000165" "24624551000194" "24921465000143"
#> [1105] "24744012000199" "24990777000109" "25005683000109" "25249439000183"
#> [1109] "26072786000146" "25369840000157" "26173550000104" "26345998000150"
#> [1113] "26735020000102" "26462693000128" "26664057000189" "26609050000164"
#> [1117] "26845702000160" "26643756000142" "26659061000159" "26982634000180"
#> [1121] "27092587000162" "27137879000174" "27157474000106" "27956690000103"
#> [1125] "28201130000101" "28442718000157" "27967244000102" "28019100000189"
#> [1129] "28594234000123" "29088092000195" "28925264000175" "28942127000149"
#> [1133] "30545324000176" "30702767000123" "29894663000189" "29845329000135"
#> [1137] "30213493000109" "29884545000190" "30219233000140" "30418658000189"
#> [1141] "30311067000108" "30788149000148" "27743642000218" "31001230000107"
#> [1145] "31468139000198" "31907982000123" "31553627000101" "31594055000109"
#> [1149] "31910308000106" "31701408000114" "31748174000160" "32372085000125"
#> [1153] "32021201000161" "32161500000100" "32431519000110" "32785497000197"
#> [1157] "32770457000171" "32803503000191" "33679797000154" "33572408000197"
#> [1161] "33839910000111" "34591574000101" "34469625000119" "35082277000195"
#> [1165] "35522178000187" "35522281000127" "35593905000105" "35502310000199"
#> [1169] "35727157000106" "35822503000127" "36243187000100" "35980592000130"
#> [1173] "36146575000164" "36729398000149" "36763716000198" "37663076000107"
#> [1177] "37592100000164" "37368315000104" "38543489000110" "38042694000100"
#> [1181] "37702340000174" "38307135000177" "38126888000186" "38482780000126"
#> [1185] "38199406000118" "38456921000136" "39393829000137" "40593395000106"
#> [1189] "39881421000104" "39580673000101" "40446476000175" "40180943000168"
#> [1193] "40760921000177" "40159947000164" "40337136000106" "40480481000102"
#> [1197] "40997635000120" "40885785000141" "42737954000121" "41508382000146"
#> [1201] "41052420000107" "41811375000119" "41335131000107" "41534746000162"
#> [1205] "41570356000148" "41886692000102" "42603513000137" "42233757000175"
#> [1209] "42310775000103" "42644220000106" "42500384000151" "42500384000151"
#> [1213] "42292007000174" "42353180000135" "41096674000119" "42937550000181"
#> [1217] "42938083000104" "42945876000150" "43335774000186" "43513237000189"
#> [1221] "43514079000181" "44067725000172" "44319688000142" "44854238000150"
#> [1225] "44841035000129" "45456117000112" "45736131000170" "47586663000101"
#> [1229] "49314049000108"