Erros robustos

Estimação consistente com a heterocedasticidade e/ou autocorrelação da matriz de covariância das estimativas de coeficientes em modelos de regressão.

utl_robErro(reg, tipo = "HC1", HAC = F)

Arguments

reg

Um objeto do tipo "lm"

tipo

String informando qual a correção da heterocedasticidade (HC1, HC2...)

HAC

Se FALSE(default), os erros serão robustos à heterocedasticidade. Se TRUE, serão robustos à heterocedasticidade e a autocorrelação.

Details

Informações adicionais sobre como usar o pacote, orientamos acessar o menu cntdd do Blog do Projeto contabiliDados: https://contabilidados.quarto.pub/. Ao acessar, fazer busca pelo nome da função utl_robErro

Contatos pelo email do Projeto contabiliDados: Email: contabilidados@ufersa.edu.br Siga-nos no Instagram: https://www.instagram.com/contabilidados @contabilidados

See also

Examples

library(cntdd)
library(sandwich)
dt <- na.omit(cntdd::dt_contabil)

estima <- lm(log(estoques) ~ log(receita), dt)

utl_robErro(estima)
#>  (Intercept) log(receita) 
#>   0.67216641   0.04465494 

utl_robErro(estima, tipo = "HC3")
#>  (Intercept) log(receita) 
#>   0.70675103   0.04708854 

utl_robErro(estima, HAC = TRUE)
#>  (Intercept) log(receita) 
#>   1.02003935   0.06664843