Tabela de média por grupos

Calcula a média de uma variável numérica baseada em grupos de duas variáveis. trata-se de uma análise tridimensional

utl_tableStat(bd, grp1, n_grp1, grp2, n_grp2, value, ...)

Arguments

bd

Um data.frame

col_Value

String informando o nome da coluna referente à variável númerica sobre a qual se pretende criar os grupos

col_Grp

String com o nome da nova coluna a ser criada com os grupos

n_grp

Número de grupos, se pretende criar grupos com o mesmo número de observações ou um vetor contendo os percentis para cada grupo. Padrão igual 2.

lbl_grp

String com o prefixo posto no início do nome do de cada grupo. Padrão igual "grp"

Examples

library(cntdd)

## Grupo 1 com os primeiros 20% das observações de menor valor, Grupo 2 com as observações de
## 20% a 80% e Grupo 3 com os 20% das observações de maior valor
utl_createGroup(mtcars, "mpg", "grp.mpg", c(0, .2, .8, 1))
#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb grp.mpg
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4   mpg_2
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4   mpg_2
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1   mpg_2
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1   mpg_2
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2   mpg_2
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1   mpg_2
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4   mpg_1
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2   mpg_3
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2   mpg_2
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4   mpg_2
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4   mpg_2
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3   mpg_2
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3   mpg_2
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3   mpg_1
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4   mpg_1
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4   mpg_1
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4   mpg_1
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1   mpg_3
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2   mpg_3
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1   mpg_3
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1   mpg_2
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2   mpg_2
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2   mpg_1
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4   mpg_1
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2   mpg_2
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1   mpg_3
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2   mpg_3
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2   mpg_3
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4   mpg_2
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6   mpg_2
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8   mpg_1
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2   mpg_2

## Grupo 1 com metade das observações e Grupo 2 com a metade restante
utl_createGroup(mtcars, "mpg", "grp.mpg", 2)
#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb grp.mpg
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4   mpg_2
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4   mpg_2
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1   mpg_2
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1   mpg_2
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2   mpg_1
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1   mpg_1
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4   mpg_1
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2   mpg_2
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2   mpg_2
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4   mpg_1
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4   mpg_1
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3   mpg_1
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3   mpg_1
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3   mpg_1
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4   mpg_1
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4   mpg_1
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4   mpg_1
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1   mpg_2
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2   mpg_2
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1   mpg_2
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1   mpg_2
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2   mpg_1
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2   mpg_1
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4   mpg_1
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2   mpg_1
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1   mpg_2
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2   mpg_2
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2   mpg_2
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4   mpg_1
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6   mpg_2
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8   mpg_1
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2   mpg_2