Cria grupos

Cria grupos baseados nos percentise de uma variável númerica

utl_createGroup(bd, col_Value, col_Grp, n_grp = 2)

Arguments

bd

Um data.frame

col_Value

String informando o nome da coluna referente à variável númerica sobre a qual se pretende criar os grupos

col_Grp

String com o nome da nova coluna a ser criada com os grupos

n_grp

Número de grupos, se pretende criar grupos com o mesmo número de observações ou um vetor contendo os percentis para cada grupo. Padrão igual 2.

lbl_grp

String com o prefixo posto no início do nome do de cada grupo. Padrão igual "grp"

See also

Examples

library(cntdd)
library(dplyr)

## Grupo 1 com os primeiros 20% das observações de menor valor, Grupo 2 com as observações de
## 20% a 80% e Grupo 3 com os 20% das observações de maior valor
utl_createGroup(mtcars, "mpg", "grp.mpg", c(0, .2, .8, 1)) %>% head()
#>                    mpg cyl disp  hp drat    wt  qsec vs am gear carb grp.mpg
#> Mazda RX4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4   mpg_2
#> Mazda RX4 Wag     21.0   6  160 110 3.90 2.875 17.02  0  1    4    4   mpg_2
#> Datsun 710        22.8   4  108  93 3.85 2.320 18.61  1  1    4    1   mpg_2
#> Hornet 4 Drive    21.4   6  258 110 3.08 3.215 19.44  1  0    3    1   mpg_2
#> Hornet Sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2   mpg_2
#> Valiant           18.1   6  225 105 2.76 3.460 20.22  1  0    3    1   mpg_2

## Grupo 1 com metade das observações e Grupo 2 com a metade restante
utl_createGroup(mtcars, "mpg", "grp.mpg", 2) %>% head()
#>                    mpg cyl disp  hp drat    wt  qsec vs am gear carb grp.mpg
#> Mazda RX4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4   mpg_2
#> Mazda RX4 Wag     21.0   6  160 110 3.90 2.875 17.02  0  1    4    4   mpg_2
#> Datsun 710        22.8   4  108  93 3.85 2.320 18.61  1  1    4    1   mpg_2
#> Hornet 4 Drive    21.4   6  258 110 3.08 3.215 19.44  1  0    3    1   mpg_2
#> Hornet Sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2   mpg_1
#> Valiant           18.1   6  225 105 2.76 3.460 20.22  1  0    3    1   mpg_1