Tabela de média por grupos

Calcula a média de uma variável numérica baseada em grupos de duas variáveis. trata-se de uma análise tridimensional

utl_tableStat(bd, grp1, n_grp1, grp2, n_grp2, value, ...)

Arguments

bd

Um data.frame

col_Value

String informando o nome da coluna referente à variável númerica sobre a qual se pretende criar os grupos

col_Grp

String com o nome da nova coluna a ser criada com os grupos

n_grp

Número de grupos, se pretende criar grupos com o mesmo número de observações ou um vetor contendo os percentis para cada grupo. Padrão igual 2.

lbl_grp

String com o prefixo posto no início do nome do de cada grupo. Padrão igual "grp"

Examples

library(cntdd)

## Valor da média das despesas operacionais de grupos de empresas baseadas no
## valor do caixa e equivalentes de caixa e nas dívidas de curto prazo.
## No exemplo, as empresas com mais dívidas de curto prazo e maiores valores
## de caixa e equivalentes de caixa são as que possuem os maiores valores
## médios de despesas operacionais ($ 1.999.101).

utl_tableStat(cntdd::dt_contabil, "cxEquiv", 3, "dividasCP", 2, "despOper", mean)
#> Error in dplyr::select_at(., .vars = value): Can't subset columns that don't exist.
#> x Column `despOper` doesn't exist.

## Por outro lado, as empresas com valores intermediários de caixa e equivalentes
## caixa (cxEquiv_2) e com maiores valores de dívidas de curto prazo são as que
## apresentam maior desvio-padrão entre os valores médios ($ 874.873).

utl_tableStat(cntdd::dt_contabil, "cxEquiv", 3, "dividasCP", 2, "despOper", sd)
#> Error in dplyr::select_at(., .vars = value): Can't subset columns that don't exist.
#> x Column `despOper` doesn't exist.